# spec/openapi.yaml
openapi: '3.0.0'
info:
   version: 1.0.0
   title: Membership portal
   license:
      name: Eclipse Public License - 2.0
      url: https://www.eclipse.org/legal/epl-2.0/
servers:
-  url: TBD - Internal
   description: Production endpoint for the membership portal data
...
paths:
...
   /form/{id}/working_groups:
      parameters:
      -  name: id
         in: path
         description: Unique ID of the membership form entry
         required: true
         schema:
            $ref: '#/components/schemas/ObjectID'
      get:
         tags:
         - Form Working Groups
         summary: Form Working Groups list
         description: Returns a list of tracked Form Working Groups for membership form
         responses:
            200:
               description: Success
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/FormWorkingGroups'
            403:
               description: Unauthorized
            500:
               description: Error while retrieving data
      post:
         tags:
         - Form Working Groups
         summary: Form Working Groups create
         description: Using the passed Form Working Groups object, creates a new object in the dataset.
         requestBody:
            description: The working group to be updated or inserted into the data set.
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/FormWorkingGroupPush'
         responses:
            200:
               description: Success
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/FormWorkingGroup'
            403:
               description: Unauthorized
   /form/{id}/working_groups/{workingGroupID}:
      parameters:
      -  name: id
         in: path
         description: Unique ID of the membership form entry
         required: true
         schema:
            $ref: '#/components/schemas/ObjectID'
      -  name: workingGroupID
         in: path
         description: Unique ID of the working group relationship within a form
         required: true
         schema:
            $ref: '#/components/schemas/ObjectID'
      get:
         tags:
         - Form Working Groups
         summary: Form Working Group
         description: Returns a working group entry that has a matching ID 
         responses:
            200:
               description: Success
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/FormWorkingGroup'
            403:
               description: Unauthorized
            500:
               description: Error while retrieving data
      put:
         tags:
         - Form Working Groups
         summary: Form Working Groups update
         description: Using the passed Form Working Groups object, updates or inserts are made to save the data.
         requestBody:
            description: The working group to be updated or inserted into the data set.
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/FormWorkingGroupPush'
         responses:
            200:
               description: Success
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/FormWorkingGroup'
            403:
               description: Unauthorized
      delete:
         tags:
         - Form Working Groups
         summary: Form Working Groups delete
         description: Remove a working group form entry with the given ID
         responses:
            204:
               description: Success
            403:
               description: Unauthorized
...
   /working_groups/{documentID}:
      parameters:
      -  name: documentID
         in: path
         description: The document ID for the working group to retrieve
         required: true
         schema:
            $ref: '#/components/schemas/ObjectID'
      get:
         tags:
         - Working Groups
         summary: Working Group
         description: Returns a working group entry that has a matching ID 
         responses:
            200:
               description: Success
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/WorkingGroup'
            500:
               description: Error while retrieving data
components:
   schemas:
      ObjectID:
         description: Unique identifier for an addressable object in the API.
         type: string
         minimum: 1
      DateTime:
         type: string
         format: datetime
         description: |
            Date string in the RFC 3339 format. Example, `1990-12-31T15:59:60-08:00`.

            More on this standard can be read at https://tools.ietf.org/html/rfc3339.
...
      Contact:
        type: object
        properties:
            id:
               $ref: '#/components/schemas/ObjectID'
            form_id:
               $ref: '#/components/schemas/ObjectID'
            first_name:
               type: string
               description: First name of the contact
            last_name:
               type: string
               description: Last name of the contact
            email:
               type: string
               description: Professional email address at which the individual can be reached at
            job_title:
               type: string
               description: Job title of the contact
            type:
               $ref: '#/components/schemas/ContactType'
...
      ContactType:
         type: string
         description: The type of contact that is being stored
         enum: [COMPANY, WORKING_GROUP, MARKETING, ACCOUNTING]
...
      FormWorkingGroups:
         type: array
         items:
            $ref: '#/components/schemas/FormWorkingGroup'
      FormWorkingGroup:
         type: object
         properties:
            id:
               $ref: '#/components/schemas/ObjectID'
            working_group:
               type: string
               description: The ID of the working group relationship to be formed
            participation_level:
               type: string
               description: The desired participation level desired within the new working group
            effective_date:
               $ref: '#/components/schemas/DateTime'
            contact:
               $ref: '#/components/schemas/Contact'
      FormWorkingGroupPush:
         type: object
         properties:
            working_group:
               type: string
               description: The ID of the working group relationship to be formed
            participation_level:
               type: string
               description: The desired participation level desired within the new working group
            effective_date:
               $ref: '#/components/schemas/DateTime'
            contact:
               $ref: '#/components/schemas/Contact'
...
      WorkingGroup:
         type: object
         properties:
            documentID:
               $ref: '#/components/schemas/ObjectID'
            name:
               type: string
            levels:
               type: object
               description: The ID of the working group relationship to be formed
               properties:
                  document_id:
                     $ref: '#/components/schemas/ObjectID'
                  level:
                     type: string
                  price:
                     type: string
                  revenue_top:
                     type: number
                  revenue_bottom:
                     type: number
                  company_size:
                     type: number